\name{plot.svm}
\alias{plot.svm}
\title{
Plot Function for "svm"
}
\description{
Visualize the solution path of regression estimate corresponding to regularization paramters.
}
\usage{
\method{plot}{svm}(x, ...)
}
\arguments{
  \item{x}{
  An object with S3 class \code{"svm"}.
  }
  \item{\dots}{
  Arguments to be passed to methods.
  }
}

\author{
Xingguo Li, Tuo Zhao, Daniel Vainsencher and Han Liu \cr
Maintainer: Xingguo Li <xingguo.leo@gmail.com>
}

\seealso{
\code{\link{stocml}} and \code{\link{stocml-package}}.
}
