\name{print.ridge}
\alias{print.ridge}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print Function for an object with S3 class \code{"ridge"}
}
\description{
Print a summary of the information about an object with S3 class \code{"ridge"}.
}
\usage{
\method{print}{ridge}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
  An object with S3 class \code{"ridge"}.
  }
  \item{\dots}{
  Arguments to be passed to methods.
  }
}
\details{
This call simply outlines the options used for computing a ridge object.
}
%\value{
%}
%\references{
%}

\author{
Xingguo Li, Tuo Zhao, Daniel Vainsencher and Han Liu \cr
Maintainer: Xingguo Li <xingguo.leo@gmail.com>
}
%\note{
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{stocml}} and \code{\link{stocml-package}}.
}
