\name{stocml-package}
\alias{stocml-package}
\docType{package}
\title{
stocml: Stochastic Machine Learning
}
\description{

}
\details{
\tabular{ll}{
Package: \tab stocml\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2015-6-29\cr
License: \tab GPL-2\cr
}
}

\author{
Xingguo Li, Tuo Zhao and Daniel Vainsencher \cr
Maintainer: Xingguo Li <xingguo.leo@gmail.com>
}

\seealso{
\code{\link{stocml}}
}
