\name{stocml-package}
\alias{stocml-package}
\docType{package}
\title{
stocml: Stochastic Machine Learning
}
\description{
Implement a family of efficient algorithms, including linear regression and linear classification, via stochastic optimization. 
}
\details{
\tabular{ll}{
Package: \tab stocml\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2015-8-15\cr
License: \tab GPL-2\cr
}
}

\author{
Xingguo Li, Tuo Zhao, Daniel Vainsencher and Han Liu \cr
Maintainer: Xingguo Li <xingguo.leo@gmail.com>
}

\seealso{
\code{\link{stocml}}
}
